solve errors once more
