+++ netlify.toml
@@ -1,7 +1,7 @@
 [build]
   command = "npm run build"
   publish = ".next"
-  functions = "netlify/functions"
+  functions = ".netlify/functions"  # Correct path if needed
 
 [[plugins]]
   package = "@netlify/plugin-nextjs"
